app/requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
blinker~=1.7.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dadata==24.10.0
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask~=2.2.5
Flask-WTF==1.2.1
geopandas==1.0.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous~=2.1.2
Jinja2~=3.1.2
MarkupSafe~=2.1.3
nest-asyncio==1.6.0
numpy~=1.26.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas~=2.1.3
plotly==5.24.1
pycparser==2.22
pyogrio==0.10.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil~=2.8.2
python-dotenv==1.0.1
pytz~=2024.1
requests==2.32.3
retrying==1.3.4
selenium==4.25.0
seleniumbase==4.32.4
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==9.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3~=2.0.3
websocket-client==1.8.0
Werkzeug~=2.2.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.20.2